requirements.txt
cryptography==2.3.1
disqus==0.0.4
Django==2.1.2
django-crispy-forms==1.7.2
django-disqus==0.5
django-formtools==2.1
django-import-export==1.1.0
django-reversion==3.0.1
gunicorn==19.9.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
mkdocs==0.16.3
mkdocs-bootswatch==1.0
mock==2.0.0
mysqlclient==1.3.13
openpyxl==2.5.9
Pillow==5.3.0
PyMySQL==0.9.2
pytz==2018.6
unicodecsv==0.14.1
uWSGI==2.0.17.1
xadmin==2.0.1
xlrd==1.1.0
xlwt==1.3.0